backend/requirements.txt
Flask
python-dotenv
psycopg2-binary
gunicorn
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.18.0
pandas==0.25.3
Pillow==6.2.1
scipy==1.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
